Release 1.7.2

+++ clpipe/get_reports.py
@@ -1,60 +1,62 @@
-import click
 from .config_json_parser import ClpipeConfigParser
 import os
 import glob
 import shutil
 from distutils.dir_util import copy_tree, remove_tree
-import sys
-import logging
-from .error_handler import exception_handler
+from tqdm import tqdm
+from .utils import add_file_handler, get_logger, resolve_fmriprep_dir
 
+STEP_NAME = "reports_fmriprep"
 
-@click.command()
-@click.option('-config_file', type=click.Path(exists=True, dir_okay=False, file_okay=True), required=True, default=None,
-              help='The configuration file for the current data processing setup.')
-@click.option('-output_name', default='Report_Archive',
-              help='Path and name of the output archive. Defaults to current working directory and "Report_Archive.zip"')
-@click.option('-debug', is_flag=True, help='Print traceback on errors.')
-def get_reports(config_file, output_name, debug):
+def get_reports(config_file, output_name, debug, clear_temp=True):
     """This command creates a zip archive of fmriprep QC reports for download."""
-    if not debug:
-        sys.excepthook = exception_handler
-        logging.basicConfig(level=logging.DEBUG)
-    else:
-        logging.basicConfig(level=logging.INFO)
-
     if config_file is None:
         raise ValueError('Please specify a configuration file.')
 
     config = ClpipeConfigParser()
     config.config_updater(config_file)
 
+    project_dir = config.config["ProjectDirectory"]
+
+    add_file_handler(os.path.join(project_dir, "logs"))
+    logger = get_logger(STEP_NAME, debug=debug)
+
     fmriprepdir = config.config['FMRIPrepOptions']['OutputDirectory']
 
-    image_dirs = [f.path for f in os.scandir(os.path.join(fmriprepdir, 'fmriprep')) if f.is_dir()]
+    logger.info(f"Generating an fMRIPrep report targeting: {fmriprepdir}")
+    logger.debug(f"Using config file: {config_file}")
 
-    for sub in [x for x in image_dirs if 'sub-' in x]:
-        logging.info(sub)
+    fmriprepdir = resolve_fmriprep_dir(fmriprepdir)
+
+    image_dirs = [f.path for f in os.scandir(fmriprepdir) if f.is_dir()]
+
+    logger.info(f"Copying figures:")
+    for sub in tqdm([x for x in image_dirs if 'sub-' in x], ascii=' #'):
         copy_tree(os.path.join(sub, 'figures'),
-                  os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp',
+                  os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp', 'fmriprep_reports',
                                os.path.basename(sub), 'figures'))
         ses_dirs = [f.path for f in os.scandir(sub) if f.is_dir()]
         for ses in [x for x in ses_dirs if 'ses-' in x]:
             if os.path.isdir(os.path.join(ses, 'figures')):
                    copy_tree(os.path.join(ses, 'figures'),
-                        os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp',
+                        os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp', 'fmriprep_reports',
                                    os.path.basename(sub),os.path.basename(ses), 'figures'))
     images = glob.glob(os.path.join(fmriprepdir, 'fmriprep', '*.html'))
 
+    logger.info(f"Copying reports...")
     for report in images:
-        logging.info(report)
         shutil.copy(report,
-                    os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp',
+                    os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp', 'fmriprep_reports',
                                  os.path.basename(report)))
 
+    logger.info(f"Creating ZIP archive...")
     shutil.make_archive(base_name=output_name,
                         root_dir=os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp'),
-                        base_dir=os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp'),
+                        base_dir='fmriprep_reports',
                         format='zip')
 
-    remove_tree(os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp'))
+    if clear_temp:
+        logger.info(f"Removing temporary directory...")
+        remove_tree(os.path.join(config.config['FMRIPrepOptions']['WorkingDirectory'], 'reports_temp'))
+
+    logger.info(f"Job finished. ZIP file created at: {output_name}")
